Use a simpler, repeated flow for d3 attributes.
Rather than sometimes adding attributes immediately and sometimes
waiting until after and then selecting and adjusting the element, try to
do things in a singular flow.  Only when an element (i.e., `container`)
is needed to be reused do you add and adjust in separate commands, and
then do as little as possible in the add step to keep the separation
clearer.

This was done to make this code a bit simpler to understand (at least
for me) and hopefully find some ways that things can be abstracted out.