Add a title attr for the labels on a count graph.
This is a bit of bleed over from our specific implementation details,
but this graph can be used where these count labels may have limited
space.  In our use of them we are hiding the overflow which means that
the full item label is not visible if it doesn't fit in the box it is
given.  By also adding a title attribute, most browsers will at least
provide a handy tooltip with the full text of the item label is
available.